Created a new records.scoped field on the records table and propogated necesssary code changes for insert, update, and delete processes.
Also modified the config.yml file and process as well as command line
arguments processing to handle the passing of a csv string of MARC
language codes for use in checking the language scope of records.

Finally, add code to the ScopeChecker.applyScopeFilter that gets
all records that have not been scoped and then loops through
the return for processing.  Loop is currently temporarily limited
to processing only 10 of the returned results.  It is also only
currently calling ScopeChecker.languageScope.

Next chunk of work is to build out ScopeChecker.languageScope, which
will complete Issue EMI-5.  Then need to add other scope tests
to the process loop.

	modified:   bin/com/carlstahmer/estc/recordimport/daemon/Conf.class
	modified:   bin/com/carlstahmer/estc/recordimport/daemon/Listener.class
	modified:   bin/com/carlstahmer/estc/recordimport/daemon/ProcessManager.class
	modified:   bin/com/carlstahmer/estc/recordimport/daemon/RecordUtils.class
	modified:   bin/com/carlstahmer/estc/recordimport/daemon/ScopeChecker.class
	modified:   bin/com/carlstahmer/estc/recordimport/daemon/SqlModel.class
	modified:   config.yml
	modified:   src/com/carlstahmer/estc/recordimport/daemon/Conf.java
	modified:   src/com/carlstahmer/estc/recordimport/daemon/Listener.java
	modified:   src/com/carlstahmer/estc/recordimport/daemon/ProcessManager.java
	modified:   src/com/carlstahmer/estc/recordimport/daemon/RecordUtils.java
	modified:   src/com/carlstahmer/estc/recordimport/daemon/ScopeChecker.java
	modified:   src/com/carlstahmer/estc/recordimport/daemon/SqlModel.java