Changes to be committed: 	modified:   src/com/carlstahmer/estc/recordimport/daemon/ExportRDF.java 	modified:   src/com/carlstahmer/estc/recordimport/daemon/ProcessManager.java 	modified:   src/com/carlstahmer/estc/recordimport/daemon/SqlModel.java
Setting up code for debugging of reading and variable assignment for new RDF model.  Changed SqlModel.java on line 604 (possibly 684, having difficulty reading my own notes) to put a limit of 1,000 records on the query for the queue.  This should run a representative sample of records.  Based on how data is put into the system, it should be a good mix of bib and holding, but I should verify this.

I also commented out code in ExportRDF.java on lines 1446-1448 and 1497-1499 that writes the actual RDF files (holding and bib respectively.).  Also commented out lines 1468 and 1502 so that records don't get marked as processed during testing.

Looked into ProcessManager.java to see if I needed to change anything there.  I must have added a space or something because it is making me check it in, but I don't think there is a substantive difference betweent this committ and the last.  If something is screwing, try a diff.